feat: se genera modelo que encapsula la logica de generacion de puntajes
