+++ controllers/solicitud_produccion.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"strconv"
-	"strings"
 
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
@@ -96,89 +95,13 @@ func (c *SolicitudProduccionController) PutResultadoSolicitud() {
 	fmt.Println("Id es: " + idStr)
 	var SolicitudProduccion map[string]interface{}
 	if err := json.Unmarshal(c.Ctx.Input.RequestBody, &SolicitudProduccion); err == nil {
-		produccionAcademica := SolicitudProduccion["ProduccionAcademica"].(map[string]interface{})
-		subTipoProduccionId := produccionAcademica["SubtipoProduccionId"].(map[string]interface{})
-		idSubtipo := subTipoProduccionId["Id"]
-		idSubtipoStr := fmt.Sprintf("%v", idSubtipo)
-		Metadatos := produccionAcademica["Metadatos"].([]interface{})
-		var autores float64
-		autores = 0
-		var valor int
-		valor = 1
-		for _, metaDatotemp := range Metadatos {
-			metaDato := metaDatotemp.(map[string]interface{})
-			metaDatoSubtipo := metaDato["MetadatoSubtipoProduccionId"].(map[string]interface{})
-			tipoMetadatoId := metaDatoSubtipo["TipoMetadatoId"].(map[string]interface{})
-			idTipoMetadato := tipoMetadatoId["Id"]
-			idTipoMetadatoStr := fmt.Sprintf("%v", idTipoMetadato)
-			idSubtipoInt, _ := strconv.Atoi(idTipoMetadatoStr)
-			if idSubtipoInt == 38 {
-				numTipoMetadatoStr := fmt.Sprintf("%v", metaDato["Valor"])
-				valor, _ = strconv.Atoi(numTipoMetadatoStr)
-			} else if idSubtipoInt == 43 {
-				numTipoMetadatoStr := fmt.Sprintf("%v", metaDato["Valor"])
-				valor, _ = strconv.Atoi(numTipoMetadatoStr)
-			} else if idSubtipoInt == 44 {
-				numTipoMetadatoStr := fmt.Sprintf("%v", metaDato["Valor"])
-				valor, _ = strconv.Atoi(numTipoMetadatoStr)
-			}
-			if idSubtipoInt == 21 {
-				numTipoMetadatoStr := fmt.Sprintf("%v", metaDato["Valor"])
-				autores, _ = strconv.ParseFloat(numTipoMetadatoStr, 64)
-			}
-		}
-		var resultado float64
-		var puntajes []map[string]interface{}
-		errProduccion := request.GetJson("http://"+beego.AppConfig.String("ProduccionAcademicaService")+"/puntaje_subtipo_produccion/?query=SubTipoProduccionId:"+idSubtipoStr, &puntajes)
-		if errProduccion == nil && fmt.Sprintf("%v", puntajes[0]["System"]) != "map[]" {
-			if puntajes[0]["Status"] != 404 && puntajes[0]["Id"] != nil {
-
-				Puntajes := puntajes[valor-1]
-
-				type Caracteristica struct {
-					Puntaje string
-				}
-				var caracteristica Caracteristica
-				json.Unmarshal([]byte(fmt.Sprintf("%v", Puntajes["Caracteristicas"])), &caracteristica)
-				puntajeStr := caracteristica.Puntaje
-				puntajeStrF := strings.ReplaceAll(puntajeStr, ",", ".")
-				puntajeInt, _ := strconv.ParseFloat(puntajeStrF, 64)
-
-				if autores <= 3 && autores > 0 {
-					resultado = puntajeInt
-					resultadoStr := strconv.FormatFloat(resultado, 'f', -1, 64)
-					SolicitudProduccion["Resultado"] = `{"Puntaje":` + resultadoStr + `}`
-				} else if autores > 3 && autores <= 5 {
-					resultado = (puntajeInt / 2)
-					resultadoStr := strconv.FormatFloat(resultado, 'f', -1, 64)
-					SolicitudProduccion["Resultado"] = `{"Puntaje":` + resultadoStr + `}`
-				} else if autores > 5 {
-					resultado = (puntajeInt / autores)
-					resultadoStr := strconv.FormatFloat(resultado, 'f', -1, 64)
-					SolicitudProduccion["Resultado"] = `{"Puntaje":` + resultadoStr + `}`
-				} else {
-					resultado = puntajeInt
-					resultadoStr := strconv.FormatFloat(resultado, 'f', -1, 64)
-					SolicitudProduccion["Resultado"] = `{"Puntaje":` + resultadoStr + `}`
-				}
-
-				c.Data["json"] = SolicitudProduccion
-
-			} else {
-				if puntajes[0]["Message"] == "Not found resource" {
-					c.Data["json"] = nil
-				} else {
-					logs.Error(puntajes)
-					c.Data["system"] = errProduccion
-					c.Abort("404")
-				}
-			}
+		if SolicitudProduccionResult, errPuntaje := models.GenerateResult(SolicitudProduccion); errPuntaje == nil {
+			c.Data["json"] = SolicitudProduccionResult
 		} else {
-			logs.Error(puntajes)
-			c.Data["system"] = errProduccion
-			c.Abort("404")
+			logs.Error(SolicitudProduccionResult)
+			c.Data["system"] = errPuntaje
+			c.Abort("400")
 		}
-
 	} else {
 		logs.Error(err)
 		c.Data["system"] = err
